backend/requirements.txt
# Core FastAPI and database
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP requests and external APIs
httpx==0.25.2
aiohttp==3.9.1

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Data processing and ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0